Updated package name of settings related to AssignmentType
Breaking change: org.instancio.assignment.* classes moved
to org.instancio.setting package
