+++ instancio-core/src/main/java/org/instancio/settings/OnSetFieldError.java
@@ -13,29 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.instancio.assignment;
+package org.instancio.settings;
 
+import org.instancio.documentation.ExperimentalApi;
 import org.instancio.exception.InstancioApiException;
 import org.instancio.internal.util.StringUtils;
-import org.instancio.settings.Keys;
 
 /**
- * A setting that specifies what should happen if a setter cannot be resolved.
+ * A setting that specifies what should happen if an error occurs
+ * when setting a field's value.
  *
- * @see Keys#ON_SET_METHOD_NOT_FOUND
+ * @see Keys#ON_SET_FIELD_ERROR
  * @since 2.1.0
  */
-public enum OnSetMethodNotFound {
+@ExperimentalApi
+public enum OnSetFieldError {
 
     /**
-     * Use field assignment as the fallback (default behaviour).
-     *
-     * @since 2.1.0
-     */
-    ASSIGN_FIELD,
-
-    /**
-     * Fail by throwing {@link InstancioApiException}.
+     * Propagate the error up by throwing {@link InstancioApiException}.
      *
      * @since 2.1.0
      */
@@ -43,6 +38,9 @@ public enum OnSetMethodNotFound {
 
     /**
      * Ignore the error and continue populating the object.
+     *
+     * <p><b>Warning:</b> an error caused by assigning an incompatible type
+     * is considered a user error and is never ignored.
      *
      * @since 2.1.0
      */
