+++ instancio-tests/feature-tests/src/test/java/org/instancio/test/features/assignmenttype/AssignmentTypeFieldTest.java
@@ -16,8 +16,8 @@
 package org.instancio.test.features.assignmenttype;
 
 import org.instancio.Instancio;
-import org.instancio.assignment.AssignmentType;
 import org.instancio.internal.util.SystemProperties;
+import org.instancio.settings.AssignmentType;
 import org.instancio.settings.Keys;
 import org.instancio.settings.Settings;
 import org.instancio.test.support.pojo.assignment.SetterStylePojo;
