+++ instancio-tests/feature-tests/src/test/java/org/instancio/test/features/assignmenttype/AssignmentTypeMethodTest.java
@@ -17,11 +17,11 @@ package org.instancio.test.features.assignmenttype;
 
 import org.instancio.Instancio;
 import org.instancio.Model;
-import org.instancio.assignment.AssignmentType;
-import org.instancio.assignment.SetterStyle;
 import org.instancio.internal.util.SystemProperties;
 import org.instancio.junit.InstancioExtension;
+import org.instancio.settings.AssignmentType;
 import org.instancio.settings.Keys;
+import org.instancio.settings.SetterStyle;
 import org.instancio.settings.Settings;
 import org.instancio.test.support.pojo.assignment.SetterStylePojo;
 import org.instancio.test.support.pojo.assignment.SetterStyleProperty;
