+++ instancio-tests/feature-tests/src/test/java/org/instancio/test/features/assignmenttype/OnSetMethodNotFoundTest.java
@@ -17,13 +17,13 @@ package org.instancio.test.features.assignmenttype;
 
 import org.instancio.Instancio;
 import org.instancio.InstancioApi;
-import org.instancio.assignment.AssignmentType;
-import org.instancio.assignment.OnSetMethodNotFound;
 import org.instancio.exception.InstancioApiException;
 import org.instancio.internal.util.SystemProperties;
 import org.instancio.junit.InstancioExtension;
 import org.instancio.junit.WithSettings;
+import org.instancio.settings.AssignmentType;
 import org.instancio.settings.Keys;
+import org.instancio.settings.OnSetMethodNotFound;
 import org.instancio.settings.Settings;
 import org.instancio.test.support.tags.Feature;
 import org.instancio.test.support.tags.FeatureTag;
