+++ instancio-tests/feature-tests/src/test/java/org/instancio/test/features/assignmenttype/OverloadedSetterTest.java
@@ -16,10 +16,10 @@
 package org.instancio.test.features.assignmenttype;
 
 import org.instancio.Instancio;
-import org.instancio.assignment.AssignmentType;
 import org.instancio.internal.util.SystemProperties;
 import org.instancio.junit.InstancioExtension;
 import org.instancio.junit.WithSettings;
+import org.instancio.settings.AssignmentType;
 import org.instancio.settings.Keys;
 import org.instancio.settings.Settings;
 import org.instancio.test.support.pojo.assignment.OverloadedSetterPojo;
