+++ instancio-tests/instancio-core-tests/src/test/java/org/external/errorhandling/MethodAssignerOnSetMethodErrorFailTest.java
@@ -16,9 +16,9 @@
 package org.external.errorhandling;
 
 import org.instancio.Instancio;
-import org.instancio.assignment.AssignmentType;
-import org.instancio.assignment.OnSetMethodError;
+import org.instancio.settings.AssignmentType;
 import org.instancio.settings.Keys;
+import org.instancio.settings.OnSetMethodError;
 import org.instancio.settings.Settings;
 import org.instancio.test.support.pojo.misc.SetterErrorPojo;
