+++ instancio-tests/java16-tests/src/test/java/org/instancio/test/java16/RecordAssignmentTypeTest.java
@@ -1,9 +1,9 @@
 package org.instancio.test.java16;
 
 import org.instancio.Instancio;
-import org.instancio.assignment.AssignmentType;
-import org.instancio.assignment.OnSetMethodError;
-import org.instancio.assignment.OnSetMethodNotFound;
+import org.instancio.settings.AssignmentType;
+import org.instancio.settings.OnSetMethodError;
+import org.instancio.settings.OnSetMethodNotFound;
 import org.instancio.junit.InstancioExtension;
 import org.instancio.settings.Keys;
 import org.instancio.settings.Settings;
